+++ packages/genesys/packages/react-components/src/Alert/Alert.tsx
@@ -8,7 +8,7 @@ import { useMergeDefaultProps } from "@peersyst/react-components-core";
 const Alert = (props: AlertProps): JSX.Element => {
     const {
         icon: iconProp = true,
-        message,
+        content,
         type,
         action,
         className,
@@ -31,7 +31,7 @@ const Alert = (props: AlertProps): JSX.Element => {
             <Row flex={1} gap={10} wrap wrapGap={10} justifyContent="space-between">
                 <Row flex={1} alignItems="flex-start" gap={10}>
                     <Row flex={0.05}>{icon}</Row>
-                    <Row flex={0.95}>{message}</Row>
+                    <Row flex={0.95}>{content}</Row>
                 </Row>
                 {action && <AlertAction className="AlertAction">{action}</AlertAction>}
             </Row>